+++ src/Control/Effect/Cut.hs
@@ -4,8 +4,6 @@ module Control.Effect.Cut
 , cutfail
 , call
 , cut
-, Branch(..)
-, branch
 , runCut
 , CutC(..)
 ) where
@@ -26,75 +24,66 @@ deriving instance Functor (Cut m)
 instance HFunctor Cut where
   hmap _ Cutfail    = Cutfail
   hmap f (Call m k) = Call (f m) k
+  {-# INLINE hmap #-}
 
 instance Effect Cut where
   handle _     _       Cutfail    = Cutfail
   handle state handler (Call m k) = Call (handler (m <$ state)) (handler . fmap k)
+  {-# INLINE handle #-}
 
 -- | Fail the current branch, and prevent backtracking within the nearest enclosing 'call' (if any).
 --
 --   Contrast with 'empty', which fails the current branch but allows backtracking.
 --
---   prop> run (runNonDetOnce (runCut (cutfail <|> pure a))) == Nothing
---   prop> run (runNonDetOnce (runCut (pure a <|> cutfail))) == Just a
+--   prop> run (runNonDet (runCut (cutfail <|> pure a))) == []
+--   prop> run (runNonDet (runCut (pure a <|> cutfail))) == [a]
 cutfail :: (Carrier sig m, Member Cut sig) => m a
 cutfail = send Cutfail
+{-# INLINE cutfail #-}
 
 -- | Delimit the effect of 'cutfail's, allowing backtracking to resume.
 --
---   prop> run (runNonDetOnce (runCut (call (cutfail <|> pure a) <|> pure b))) == Just b
+--   prop> run (runNonDet (runCut (call (cutfail <|> pure a) <|> pure b))) == [b]
 call :: (Carrier sig m, Member Cut sig) => m a -> m a
 call m = send (Call m ret)
+{-# INLINE call #-}
 
 -- | Commit to the current branch, preventing backtracking within the nearest enclosing 'call' (if any) on failure.
 --
---   prop> run (runNonDetOnce (runCut (cut *> pure a <|> pure b))) == Just a
---   prop> run (runNonDetOnce (runCut (cut *> empty <|> pure a))) == Nothing
+--   prop> run (runNonDet (runCut (pure a <|> cut *> pure b))) == [a, b]
+--   prop> run (runNonDet (runCut (cut *> pure a <|> pure b))) == [a]
+--   prop> run (runNonDet (runCut (cut *> empty <|> pure a))) == []
 cut :: (Alternative m, Carrier sig m, Member Cut sig) => m ()
 cut = pure () <|> cutfail
-
-
--- | The result of a nondeterministic branch of a computation.
-data Branch a
-  = Prune
-  | None
-  | Some a
-  deriving (Eq, Functor, Ord, Show)
-
--- | Case analysis for 'Branch', taking a value to use for 'Prune', a value to use for 'None', and a function to apply to the contents of 'Some'.
---
---   prop> branch Prune None Some a == a
---   prop> branch a b (applyFun f) Prune == a
---   prop> branch a b (applyFun f) None == b
---   prop> branch a b (applyFun f) (Some c) == applyFun f c
-branch :: a -> a -> (b -> a) -> Branch b -> a
-branch a _ _ Prune    = a
-branch _ a _ None     = a
-branch _ _ f (Some a) = f a
+{-# INLINE cut #-}
 
 
 -- | Run a 'Cut' effect within an underlying 'Alternative' instance (typically 'Eff' carrying a 'NonDet' effect).
 --
 --   prop> run (runNonDetOnce (runCut (pure a))) == Just a
 runCut :: (Alternative m, Carrier sig m, Effect sig, Monad m) => Eff (CutC m) a -> m a
-runCut = (>>= branch empty empty pure) . runCutC . interpret
+runCut = (>>= runBranch (const empty)) . runCutC . interpret
 
-newtype CutC m a = CutC { runCutC :: m (Branch a) }
+newtype CutC m a = CutC { runCutC :: m (Branch m Bool a) }
 
 instance (Alternative m, Carrier sig m, Effect sig, Monad m) => Carrier (Cut :+: NonDet :+: sig) (CutC m) where
-  ret = CutC . ret . Some
+  ret = CutC . ret . Pure
+  {-# INLINE ret #-}
+
   eff = CutC . handleSum (handleSum
-    (eff . handle (Some ()) (branch (ret Prune) (ret None) runCutC))
+    (eff . handle (Pure ()) (bindBranch (ret (None False)) runCutC))
     (\case
-      Empty    -> ret None
-      Choose k -> runCutC (k True) >>= branch (ret Prune) (runCutC (k False)) (\ a -> ret (Some a) <|> runCutC (k False))))
+      Empty    -> ret (None True)
+      Choose k -> runCutC (k True) >>= branch (\ e -> if e then runCutC (k False) else ret (None False)) (\ a -> ret (Alt (ret a) (runCutC (k False) >>= runBranch (const empty)))) (fmap ret . Alt)))
     (\case
-      Cutfail  -> ret Prune
-      Call m k -> runCutC m >>= branch (ret None) (ret None) (runCutC . k))
+      Cutfail  -> ret (None False)
+      Call m k -> runCutC m >>= bindBranch (ret (None True)) (runCutC . k))
+    where bindBranch :: (Alternative m, Carrier sig m, Monad m) => m (Branch m Bool a) -> (b -> m (Branch m Bool a)) -> Branch m Bool b -> m (Branch m Bool a)
+          bindBranch cut bind = branch (\ e -> if e then ret (None True) else cut) bind (\ a b -> ret (Alt (a >>= bind >>= runBranch (const empty)) (b >>= bind >>= runBranch (const empty))))
+  {-# INLINE eff #-}
 
 
 -- $setup
 -- >>> :seti -XFlexibleContexts
 -- >>> import Test.QuickCheck
 -- >>> import Control.Effect.Void
--- >>> instance Arbitrary a => Arbitrary (Branch a) where arbitrary = frequency [(1, pure Prune), (1, pure None), (3, Some <$> arbitrary)] ; shrink b = case b of { Some a -> Prune : None : map Some (shrink a) ; None -> [Prune] ; Prune -> [] }