+++ ChangeLog.md
@@ -1,6 +1,8 @@
 # 0.1.2.0
 
 - Adds a `Cut` effect which adds committed choice to nondeterminism.
+- Adds a `Cull` effect which adds pruning to nondeterminism.
+- Adds an example of using `NonDet`, `Cut`, and a character parser effect to define parsers.
 - Fixes the table of contents links in the README.
 
 # 0.1.1.0